Split out an abstract base class for the strategy.
