+++ src/python/pants/backend/jvm/tasks/jvm_compile/jvm_compile.py
@@ -11,7 +11,7 @@ import sys
 from collections import defaultdict
 
 from pants.backend.core.tasks.group_task import GroupMember
-from pants.backend.jvm.tasks.jvm_compile.jvm_compile_strategy import JvmCompileStrategy
+from pants.backend.jvm.tasks.jvm_compile.jvm_compile_global_strategy import JvmCompileGlobalStrategy
 from pants.backend.jvm.tasks.jvm_compile.jvm_dependency_analyzer import JvmDependencyAnalyzer
 from pants.backend.jvm.tasks.jvm_compile.jvm_fingerprint_strategy import JvmFingerprintStrategy
 from pants.backend.jvm.tasks.nailgun_task import NailgunTaskBase
@@ -203,7 +203,7 @@ class JvmCompile(NailgunTaskBase, GroupMember):
     self._sources_by_target = None
 
     # The compile strategy to use for analysis and classfile placement.
-    self._strategy = JvmCompileStrategy(self.context,
+    self._strategy = JvmCompileGlobalStrategy(self.context,
                                         self.get_options(),
                                         self.workdir,
                                         self.create_analysis_tools())
